+++ README.md
@@ -20,7 +20,7 @@
 - 구현
   - [ ] 패키지 구조 변경 (컨트롤러, 도메인)
   - [x] 변경되지 않는 필드 final로 선언
-  - [ ] RacingCar의 init 메서드를 생성자로 변경
+  - [x] RacingCar의 init 메서드를 생성자로 변경
   - [ ] 랜덤 숫자 생성 기능을 따로 클래스로 분리
   - [ ] 우승자 판별 메서드 stream을 이용해 리팩토링
   - [ ] 클래스 내부에서만 사용하는 메서드는 private으로 선언
