+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -1,23 +1,15 @@
 package racingcar.controller;
 
-import racingcar.controller.domain.RacingGame;
+import racingcar.domain.RacingGame;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 public class RacingGameController {
 
-    private RacingGame racingGame;
-    private InputView InputView;
-    private OutputView outputView;
-
-    public RacingGameController() {
-        this.racingGame = new RacingGame();
-        this.InputView = new InputView();
-        this.outputView = new OutputView();
-    }
-
     public void start() {
-        racingGame.init(InputView.inputCarNames(),InputView.inputFinalTurn());
+        InputView inputView = new InputView();
+        OutputView outputView = new OutputView();
+        RacingGame racingGame = new RacingGame(inputView.inputCarNames(),inputView.inputFinalTurn());
         outputView.printGameResultMessage();
         outputView.printTurnResult(racingGame.getCars());
         while(!racingGame.isFinished()) {
@@ -25,4 +17,5 @@ public class RacingGameController {
         }
         outputView.printWinners(racingGame.judgeWinners());
     }
+
 }
